Updated template and fixed a bug.
There was a but in dataformat.py.  When it used Meta.ordering, it tried to get the field using the unmodified string, even if it had a "-" prepended to reverse the ordering.
Given an updated understanding of the relationships between PeakData and PeakDataLabel, I reverted the split_rows change.  There is a 1:1 relationship between a PeakData record and a single labeling state, so row splitting is unnecessary.
For some reason, my database was missing the PeakDataLabel table, and the only thing that fixed it was to blow away and remake migrations.

Files checked in: DataRepo/formats/dataformat.py DataRepo/formats/peakdata_dataformat.py DataRepo/migrations/0001_initial.py DataRepo/migrations/0002_tracer_group_name.py DataRepo/templates/DataRepo/search/results/peakdata.html